[anthony.yup] 방탈출 인증 관리 1+2 단계 (#33)
* feat(auth): 토큰 발급 기능 구현

* feat(auth): 내 정보 조회 기능 구현

* feat(auth): Auth 테스트 구현

* feat(auth): interceptor 구현

* feat(auth): resolver 구현

* feat(auth): 예약 삭제 구현

* feat(auth): reservation 테스트 구현

* feat(auth): exceptionHandler 제거

* refactor: 헤더에서 토큰 가져오는 작업 분리

* refactor: spring bean으로 등록하여 di 활용

* refactor: 주석 처리한 코드 삭제

* refactor: 패스워드 비교를 member 도메인 로직을 사용하도록 변경

* fix: 기존에 삭제했던 전체 Exception 핸들러 대신 NullPointerException 핸들러를 새로 등록

* fix: reservation 도메인 필드 이름 변경 name -> username

* test: 테스트 케이스 데이터 상수화

* fix: 예약 시 예약자 정보를 직접 넣는 것이 아닌 로그인 된 정보를 통해 가져오도록 수정

* refactor: 함수 이름 변경 JwtTokenProvider getPrincipal -> getUsername

* fix: 다른 사람의 예약 취소 시 Forbidden 을 반환하도록 수정

* test: 비로그인 상태에서 예약 취소하는 경우 테스트 추가

* refactor: 예약자 확인 로직을 reservation 도메인에서 할 수 있도록 변경

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,18 +1,21 @@
 package nextstep.reservation;
 
+import java.util.List;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.ForbiddenException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 public class ReservationService {
+
     public final ReservationDao reservationDao;
+
     public final ThemeDao themeDao;
+
     public final ScheduleDao scheduleDao;
 
     public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
@@ -21,7 +24,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, String username) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,7 +37,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                username
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +52,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String username) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if (!reservation.isOwner(username)) {
+            throw new ForbiddenException();
         }
 
         reservationDao.deleteById(id);
